refactor notification -> profile
